+++ tests/src/python/CMakeLists.txt
@@ -90,7 +90,6 @@ ADD_PYTHON_TEST(PyQgsGeometryAvoidIntersections test_qgsgeometry_avoid_intersect
 ADD_PYTHON_TEST(PyQgsGeometryGeneratorSymbolLayer test_qgsgeometrygeneratorsymbollayer.py)
 ADD_PYTHON_TEST(PyQgsGeometryTest test_qgsgeometry.py)
 ADD_PYTHON_TEST(PyQgsGeometryValidator test_qgsgeometryvalidator.py)
-ADD_PYTHON_TEST(PyQgsGpkgProvider test_provider_gpkg.py)
 ADD_PYTHON_TEST(PyQgsGraduatedSymbolRenderer test_qgsgraduatedsymbolrenderer.py)
 ADD_PYTHON_TEST(PyQgsHashLineSymbolLayer test_qgshashlinesymbollayer.py)
 ADD_PYTHON_TEST(PyQgsHighlight test_qgshighlight.py)
